whisper/requirements.txt
boto3
ffmpeg
openai-whisper
mysql-connector
urllib3